pyproject.toml
[project]
name = "ai-lesson11-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.2",
    "langfuse>=3.10.0",
    "pillow>=12.0.0",
    "pydantic-ai-slim[logfire,openai]>=1.18.0",
    "pymupdf>=1.26.6",
    "pymupdf-layout>=1.26.6",
    "pymupdf4llm>=0.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
